requirements.txt
djangorestframework
python-decouple
psycopg2-binary
djangorestframework-simplejwt
django-allauth
requests 
cryptography
pillow
django-filter
pytils